@@ -31,7 +31,7 @@
 
 	"require": {
 
-		"php": "^5.6 || ^7.0",
+		"php": "^7.1 || ^8.0",
 
 		"ext-gd": "*",
 		"ext-mbstring": "*",
@@ -45,7 +45,7 @@
 
 	"require-dev": {
 
-		"phpunit/phpunit": "^5.0",
+		"phpunit/phpunit": "^9.0",
 		"mockery/mockery": "^1.3.0",
 		"squizlabs/php_codesniffer": "^3.5.0",
 		"tracy/tracy": "^2.4",
