@@ -2,7 +2,9 @@
 
 namespace Issues;
 
-class Issue800Test extends \PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class Issue800Test extends TestCase
 {
 
 	public function testNoNoticeWithNestedTablesAndBorders()
