@@ -2,7 +2,9 @@
 
 namespace Issues;
 
-class Issue814Test extends \PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class Issue814Test extends TestCase
 {
 
 	public function testNoNoticeWithBorderColumns()
