@@ -3,7 +3,9 @@
 
 namespace Mpdf;
 
-class CacheTest extends \PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class CacheTest extends TestCase
 {
 	protected $basePath;
 	protected $oldTmpMode;
@@ -20,7 +22,7 @@ class CacheTest extends \PHPUnit_Framework_TestCase
 		return $this->basePath . $relativeToRoot;
 	}
 
-	protected function setUp()
+	protected function setUp(): void
 	{
 		parent::setUp();
 
@@ -29,7 +31,7 @@ class CacheTest extends \PHPUnit_Framework_TestCase
 		chmod($dir, 0777);
 	}
 
-	protected function tearDown()
+	protected function tearDown(): void
 	{
 		chmod($this->path("tmp"), $this->oldTmpMode);
 
