@@ -5,7 +5,7 @@ namespace Issues;
 class Issue732Test extends \Mpdf\BaseMpdfTest
 {
 
-	public function setUp()
+	public function setUp(): void
 	{
 		parent::setUp();
 
