@@ -9,7 +9,7 @@ class Issue761Test extends \Mpdf\BaseMpdfTest
 	{
 		$this->mpdf->WriteHTML('<div style="font-size: .8cm">This text shall be big</div>');
 		$this->mpdf->Close();
-		$this->assertContains('/F1 22.677 Tf', $this->mpdf->pages[1]);
+		$this->assertStringContainsString('/F1 22.677 Tf', $this->mpdf->pages[1]);
 	}
 
 }
