@@ -25,11 +25,11 @@ class Issue897Test extends \Mpdf\BaseMpdfTest
 
 		$this->mpdf->Close();
 
-		$this->assertRegexp('/0.000 0.502 0.000 rg/', $this->mpdf->pages[1]);
-		$this->assertRegexp('/0.000 0.502 0.000 RG/', $this->mpdf->pages[1]);
+		$this->assertMatchesRegularExpression('/0.000 0.502 0.000 rg/', $this->mpdf->pages[1]);
+		$this->assertMatchesRegularExpression('/0.000 0.502 0.000 RG/', $this->mpdf->pages[1]);
 
-		$this->assertNotRegexp('/0.000 0.502 0.000 rg/', $this->mpdf->pages[2]);
-		$this->assertNotRegexp('/0.000 0.502 0.000 RG/', $this->mpdf->pages[2]);
+		$this->assertDoesNotMatchRegularExpression('/0.000 0.502 0.000 rg/', $this->mpdf->pages[2]);
+		$this->assertDoesNotMatchRegularExpression('/0.000 0.502 0.000 RG/', $this->mpdf->pages[2]);
 	}
 
 }
