@@ -7,7 +7,7 @@ use Mpdf\Mpdf;
 class Issue905Test extends \Mpdf\BaseMpdfTest
 {
 
-	public function setUp()
+	public function setUp(): void
 	{
 		parent::setUp();
 
