@@ -1,19 +1,21 @@
 <?php
 namespace Mpdf\Barcode;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * @group unit
  */
-class Code128Test extends \PHPUnit_Framework_TestCase
+class Code128Test extends TestCase
 {
 
 	public function testInit()
 	{
 		$barcode = new Code128('103 33 99   12  ', 'RAW');
 		$array = $barcode->getData();
-		$this->assertInternalType('array', $array);
+		$this->assertIsArray($array);
 		$this->assertArrayHasKey('bcode', $array);
-		$this->assertInternalType('array', $array['bcode']);
+		$this->assertIsArray($array['bcode']);
 	}
 
 	public function invalidCodeProvider()
@@ -35,10 +37,10 @@ class Code128Test extends \PHPUnit_Framework_TestCase
 
 	/**
 	 * @dataProvider invalidCodeProvider
-	 * @expectedException \Mpdf\Barcode\BarcodeException
 	 */
 	public function testInvalidCode($SubType, $code)
 	{
+		$this->expectException(\Mpdf\Barcode\BarcodeException::class);
 		new Code128($code, $SubType);
 	}
 
