@@ -3,8 +3,9 @@
 namespace Mpdf;
 
 use Mpdf\Mpdf;
+use PHPUnit\Framework\TestCase;
 
-abstract class BaseMpdfTest extends \PHPUnit_Framework_TestCase
+abstract class BaseMpdfTest extends TestCase
 {
 
 	/**
@@ -12,14 +13,14 @@ abstract class BaseMpdfTest extends \PHPUnit_Framework_TestCase
 	 */
 	protected $mpdf;
 
-	protected function setUp()
+	protected function setUp(): void
 	{
 		parent::setUp();
 
 		$this->mpdf = new Mpdf(['mode' => 'c']);
 	}
 
-	protected function tearDown()
+	protected function tearDown(): void
 	{
 		parent::tearDown();
 
