@@ -4,8 +4,9 @@ namespace Mpdf\Fonts;
 
 use Mpdf\Mpdf;
 use Mpdf\Cache;
+use PHPUnit\Framework\TestCase;
 
-class FontCacheTest extends \PHPUnit_Framework_TestCase
+class FontCacheTest extends TestCase
 {
 
 	/**
@@ -18,7 +19,7 @@ class FontCacheTest extends \PHPUnit_Framework_TestCase
 	 */
 	private $fontCache;
 
-	protected function setUp()
+	protected function setUp(): void
 	{
 		parent::setUp();
 
