@@ -3,8 +3,9 @@
 namespace Mpdf\Fonts;
 
 use Mockery;
+use PHPUnit\Framework\TestCase;
 
-class MetricsGeneratorTest extends \PHPUnit_Framework_TestCase
+class MetricsGeneratorTest extends TestCase
 {
 
 	/**
@@ -17,7 +18,7 @@ class MetricsGeneratorTest extends \PHPUnit_Framework_TestCase
 	 */
 	private $fontCache;
 
-	protected function setUp()
+	protected function setUp(): void
 	{
 		parent::setUp();
 
@@ -25,7 +26,7 @@ class MetricsGeneratorTest extends \PHPUnit_Framework_TestCase
 		$this->generator = new MetricsGenerator($this->fontCache, 'win');
 	}
 
-	protected function tearDown()
+	protected function tearDown(): void
 	{
 		parent::tearDown();
 
