@@ -10,8 +10,9 @@ use Mpdf\Language\ScriptToLanguage;
 use Mpdf\Mpdf;
 use Mpdf\Otl;
 use Mpdf\SizeConverter;
+use PHPUnit\Framework\TestCase;
 
-class SvgTest extends \PHPUnit_Framework_TestCase
+class SvgTest extends TestCase
 {
 
 	/**
@@ -23,7 +24,7 @@ class SvgTest extends \PHPUnit_Framework_TestCase
 
 	private $colorConverter;
 
-	protected function setUp()
+	protected function setUp(): void
 	{
 		parent::setUp();
 
@@ -54,7 +55,7 @@ class SvgTest extends \PHPUnit_Framework_TestCase
 		);
 	}
 
-	protected function tearDown()
+	protected function tearDown(): void
 	{
 		parent::tearDown();
 
