@@ -3,8 +3,9 @@
 namespace Mpdf\Pdf;
 
 use Mockery;
+use PHPUnit\Framework\TestCase;
 
-class ProtectionTest extends \PHPUnit_Framework_TestCase
+class ProtectionTest extends TestCase
 {
 
 	/**
@@ -17,7 +18,7 @@ class ProtectionTest extends \PHPUnit_Framework_TestCase
 	 */
 	private $generator;
 
-	protected function setUp()
+	protected function setUp(): void
 	{
 		/** @var \Mpdf\Pdf\Protection\UniqidGenerator $generator */
 		$this->generator = Mockery::mock('Mpdf\Pdf\Protection\UniqidGenerator');
@@ -25,7 +26,7 @@ class ProtectionTest extends \PHPUnit_Framework_TestCase
 		$this->protection = new Protection($this->generator);
 	}
 
-	protected function tearDown()
+	protected function tearDown(): void
 	{
 		parent::tearDown();
 
