@@ -5,8 +5,9 @@ namespace Mpdf;
 use Mpdf\Pdf\Protection;
 use Mpdf\Pdf\Protection\UniqidGenerator;
 use Mpdf\Writer\BaseWriter;
+use PHPUnit\Framework\TestCase;
 
-class SubstitutionTest extends \PHPUnit_Framework_TestCase
+class SubstitutionTest extends TestCase
 {
 
 	/**
@@ -14,7 +15,7 @@ class SubstitutionTest extends \PHPUnit_Framework_TestCase
 	 */
 	private $mpdf;
 
-	protected function setUp()
+	protected function setUp(): void
 	{
 		parent::setUp();
 
