@@ -6,8 +6,9 @@ use Mockery;
 use Mpdf\Pdf\Protection;
 use Mpdf\Pdf\Protection\UniqidGenerator;
 use Mpdf\Writer\BaseWriter;
+use PHPUnit\Framework\TestCase;
 
-class TocNumbering extends \PHPUnit_Framework_TestCase
+class TocNumbering extends TestCase
 {
 
 	/**
@@ -15,12 +16,12 @@ class TocNumbering extends \PHPUnit_Framework_TestCase
 	 */
 	private $mpdf;
 
-	protected function setUp()
+	protected function setUp(): void
 	{
 		$this->mpdf = new Mpdf();
 	}
 
-	protected function tearDown()
+	protected function tearDown(): void
 	{
 		parent::tearDown();
 
@@ -64,7 +65,7 @@ class TocNumbering extends \PHPUnit_Framework_TestCase
 
 		$this->mpdf->Close();
 
-		$this->assertContains($this->getPattern(3), $this->mpdf->pages[2]);
+		$this->assertStringContainsString($this->getPattern(3), $this->mpdf->pages[2]);
 	}
 
 	/**
@@ -129,9 +130,9 @@ class TocNumbering extends \PHPUnit_Framework_TestCase
 
 		$this->mpdf->Close();
 
-		$this->assertContains($this->getPattern(7), $this->mpdf->pages[1]);
-		$this->assertContains($this->getPattern(8), $this->mpdf->pages[4]);
-		$this->assertContains($this->getPattern(9), $this->mpdf->pages[6]);
+		$this->assertStringContainsString($this->getPattern(7), $this->mpdf->pages[1]);
+		$this->assertStringContainsString($this->getPattern(8), $this->mpdf->pages[4]);
+		$this->assertStringContainsString($this->getPattern(9), $this->mpdf->pages[6]);
 	}
 
 	public function testTocAlternateSymbols()
@@ -193,11 +194,11 @@ class TocNumbering extends \PHPUnit_Framework_TestCase
 
 		$this->mpdf->Close();
 
-		$this->assertContains($this->getPattern('VII', 'q 0.000 0.000 0.000 rg  0 Tr BT 540.165 784.480 Td  (%s) Tj ET Q'), $this->mpdf->pages[1]);
+		$this->assertStringContainsString($this->getPattern('VII', 'q 0.000 0.000 0.000 rg  0 Tr BT 540.165 784.480 Td  (%s) Tj ET Q'), $this->mpdf->pages[1]);
 
-		$this->assertContains($this->getPattern('VIII', 'q 0.000 0.000 0.000 rg  0 Tr BT 537.250 784.480 Td  (%s) Tj ET Q'), $this->mpdf->pages[4]);
+		$this->assertStringContainsString($this->getPattern('VIII', 'q 0.000 0.000 0.000 rg  0 Tr BT 537.250 784.480 Td  (%s) Tj ET Q'), $this->mpdf->pages[4]);
 
-		$this->assertContains($this->getPattern('IX', 'q 0.000 0.000 0.000 rg  0 Tr BT 543.069 784.480 Td  (%s) Tj ET Q'), $this->mpdf->pages[6]);
+		$this->assertStringContainsString($this->getPattern('IX', 'q 0.000 0.000 0.000 rg  0 Tr BT 543.069 784.480 Td  (%s) Tj ET Q'), $this->mpdf->pages[6]);
 	}
 
 	public function testTocNumberSuppression()
@@ -232,7 +233,7 @@ class TocNumbering extends \PHPUnit_Framework_TestCase
 
 		$this->mpdf->Close();
 
-		$this->assertContains($this->getPattern('6', 'q 0.000 0.000 0.000 rg  0 Tr BT 546.468 741.642 Td  (%s) Tj ET Q'), $this->mpdf->pages[2]);
+		$this->assertStringContainsString($this->getPattern('6', 'q 0.000 0.000 0.000 rg  0 Tr BT 546.468 741.642 Td  (%s) Tj ET Q'), $this->mpdf->pages[2]);
 	}
 
 	public function testTocNumberWithCustomNumberStylingOnTocPage()
@@ -266,7 +267,7 @@ class TocNumbering extends \PHPUnit_Framework_TestCase
 
 		$this->mpdf->Close();
 
-		$this->assertContains($this->getPattern('5', 'q 0.000 0.000 0.000 rg  0 Tr BT 546.468 767.980 Td  (%s) Tj ET Q'), $this->mpdf->pages[2]);
+		$this->assertStringContainsString($this->getPattern('5', 'q 0.000 0.000 0.000 rg  0 Tr BT 546.468 767.980 Td  (%s) Tj ET Q'), $this->mpdf->pages[2]);
 	}
 
 	protected function getPattern(
